feat: winston2+winston3 support wip  tests broken
