+++ package.json
@@ -75,7 +75,9 @@
   "dependencies": {
     "@google-cloud/logging": "^3.0.2",
     "@sindresorhus/is": "^0.10.0",
-    "lodash.mapvalues": "^4.6.0"
+    "lodash.mapvalues": "^4.6.0",
+    "logform": "^1.9.1",
+    "winston-transport": "^4.2.0"
   },
   "devDependencies": {
     "@google-cloud/common": "^0.24.0",
@@ -90,8 +92,8 @@
     "@types/pify": "^3.0.2",
     "@types/proxyquire": "^1.3.28",
     "@types/rimraf": "^2.0.2",
+    "@types/semver": "^5.5.0",
     "@types/tmp": "0.0.33",
-    "@types/winston": "^2.3.9",
     "codecov": "^3.0.2",
     "cpy-cli": "^2.0.0",
     "delay": "^4.0.0",
@@ -120,6 +122,6 @@
     "winston": "^2.2.0"
   },
   "peerDependencies": {
-    "winston": "^2.x"
+    "winston": "^3.x"
   }
 }
